requirements.txt
streamlit>=1.28.0
slack-sdk>=3.23.0
pandas>=2.0.0
plotly>=5.17.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2023.3
streamlit-aggrid>=0.3.4
streamlit-option-menu>=0.3.6
snowflake-connector-python>=3.0.0
